Replace vector backend with a trait and glam

+++ src/hit.rs
@@ -10,7 +10,7 @@ pub struct HitRecord<'a> {
     pub p: Point3,
     pub normal: Vec3,
     pub mat: &'a dyn Scatter,
-    pub t: f64,
+    pub t: f32,
     pub front_face: bool,
 }
 
@@ -26,13 +26,13 @@ impl<'a> HitRecord<'a> {
 }
 
 pub trait Hit: Send + Sync {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord>;
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord>;
 }
 
 pub type World = Vec<Box<dyn Hit>>;
 
 impl Hit for World {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let mut temp_rec = None;
         let mut closest_so_far = t_max;
