+++ src/ray.rs
@@ -24,7 +24,7 @@ impl Ray {
         self.dir
     }
 
-    pub fn at(&self, t: f64) -> Point3 {
+    pub fn at(&self, t: f32) -> Point3 {
         self.orig + t * self.dir
     }
 }
