+++ src/sphere.rs
@@ -9,12 +9,12 @@ use super::vec::{Point3, Vec3};
 
 pub struct Sphere {
     center: Point3,
-    radius: f64,
+    radius: f32,
     mat: Arc<dyn Scatter>,
 }
 
 impl Sphere {
-    pub fn new(center: Point3, radius: f64, mat: Arc<dyn Scatter>) -> Self {
+    pub fn new(center: Point3, radius: f32, mat: Arc<dyn Scatter>) -> Self {
         Self {
             center,
             radius,
@@ -24,7 +24,7 @@ impl Sphere {
 }
 
 impl Hit for Sphere {
-    fn hit(&self, r: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, r: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let oc = r.origin() - self.center;
         let a = r.direction().length().powi(2);
         let half_b = oc.dot(r.direction());
